Added users as own test
