+++ .yamllint
@@ -1,6 +1,11 @@
----
 extends: default
+
 rules:
-  line-length:
-    max: 120
+  braces:
+    max-spaces-inside: 1
+    level: error
+  brackets:
+    max-spaces-inside: 1
+    level: error
+  line-length: disable
   truthy: disable
