+++ tasks/main.yml
@@ -124,6 +124,13 @@
     line: '%operators ALL=(ALL) NOPASSWD: ALL'
     validate: visudo -cf %s
 
+- name: Remove inactive users from sudoers
+  file: >-
+    dest=/etc/sudoers.d/{{ item.username }}
+    state=absent
+  when: not (item.active | default(false))
+  with_items: "{{ jumpbox_operators }}"
+
 - name: Configure dsh
   template: src=etc_dsh_dsh.conf.j2 dest=/etc/dsh/dsh.conf mode=0644 owner=root group=root
